Improve WASM compat
